Refactor/zig 12 (#2)
* refactor: update to zig 12

* fix: windows build

* refactor: windows prebuilt includes and libs

* fix: macos build

* fix: music loading on windows

+++ zig-src/src/animation.zig
@@ -10,6 +10,6 @@ pub const TaggedAnimation = union(enum) {
 
 pub const Animation = struct {
     duration: u32,
-    t0: usize,
+    t0: u64,
     anim_type: AnimationType,
 };
